Refactor frontend
- Restructure file tree
- Remove old boiler plate

jira: FRON-14
