+++ ubc-scheduler/src/Definitions/Types/ActionTypes/CommonActionTypes.ts
@@ -1,0 +1,8 @@
+import { ICourse } from "../../Interfaces/CommonInterfaces"
+
+export type CourseAction = {
+    type: string, 
+    courses?: ICourse[]
+}
+  
+export type DispatchTypeCourses = (args: CourseAction) => CourseAction
