+++ ubc-scheduler/src/Definitions/Types/StateTypes/CommonStateTypes.ts
@@ -1,0 +1,10 @@
+import { ICourse } from "../../Interfaces/CommonInterfaces"
+
+export interface ApplicationState {
+    courseState: CourseState | undefined
+}
+
+export type CourseState = {
+    courses?: ICourse[],
+    loading: boolean
+}
