+++ ubc-scheduler/src/Definitions/actionTypes.ts
@@ -1,6 +1,3 @@
-export const ADD_ARTICLE = "ADD_ARTICLE";
-export const REMOVE_ARTICLE = "REMOVE_ARTICLE";
-
 export const ADD_COURSE = "ADD_COURSE";
 export const GET_COURSES = "GET_COURSE";
 export const COURSES_LOADING = "COURSE_LOADING";
