--- ubc-scheduler/src/store/actions/actionCreators.ts
@@ -1,31 +1,0 @@
-import { 
-  IArticle,
-  ArticleAction,
-  DispatchType
-} from "../../types/commonTypes"
-import * as actionTypes from "../actionTypes"
-
-export function addArticle(article: IArticle) {
-  const action: ArticleAction = {
-    type: actionTypes.ADD_ARTICLE,
-    article,
-  }
-
-  return simulateHttpRequest(action)
-}
-
-export function removeArticle(article: IArticle) {
-  const action: ArticleAction = {
-    type: actionTypes.REMOVE_ARTICLE,
-    article,
-  }
-  return simulateHttpRequest(action)
-}
-
-export function simulateHttpRequest(action: ArticleAction) {
-  return (dispatch: DispatchType) => {
-    setTimeout(() => {
-      dispatch(action)
-    }, 500)
-  }
-}
