--- ubc-scheduler/src/store/reducers/articleReducer.ts
@@ -1,53 +1,0 @@
-import * as actionTypes from "../actionTypes";
-import {
-  ArticleState,
-  ArticleAction,
-  IArticle
-} from "../../types/commonTypes";
-
-export const initialArticleState: ArticleState = {
-  articles: [
-    {
-      id: 1,
-      title: "post 1",
-      body:
-        "Quisque cursus, metus vitae pharetra Nam libero tempore, cum soluta nobis est eligendi",
-    },
-    {
-      id: 2,
-      title: "post 2",
-      body:
-        "Harum quidem rerum facilis est et expedita distinctio quas molestias excepturi sint",
-    },
-  ],
-}
-
-const articleReducer = (
-    state: ArticleState = initialArticleState,
-    action: ArticleAction
-  ): ArticleState => {
-    switch (action.type) {
-      case actionTypes.ADD_ARTICLE:
-        const newArticle: IArticle = {
-          id: Math.random(), // not really unique
-          title: action.article.title,
-          body: action.article.body,
-        }
-        return {
-          ...state,
-          articles: state.articles.concat(newArticle),
-        }
-      case actionTypes.REMOVE_ARTICLE:
-        const updatedArticles: IArticle[] = state.articles.filter(
-          article => article.id !== action.article.id
-        )
-        return {
-          ...state,
-          articles: updatedArticles,
-        }
-      default: 
-        return state;
-    }
-  }
-  
-  export default articleReducer;
