+++ ubc-scheduler/src/store/reducers/index.ts
@@ -1,9 +1,7 @@
 import {combineReducers} from 'redux';
-import { ApplicationState } from '../../types/commonTypes';
-import articleReducer from './articleReducer';
+import { ApplicationState } from '../../Definitions/Types/StateTypes/CommonStateTypes';
 import courseReducer from './courseReducer';
 
 export default combineReducers<ApplicationState>({
-    articleState: articleReducer,
     courseState: courseReducer
 });
